+++ src/style.css
@@ -65,6 +65,7 @@
     align-items: center;
     justify-content: center;
     text-decoration: none;
+    border: none;
 }
 
 /*REGISTER*/
